+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - New exception `QuantityError` and the associated procedure `RaiseQuantityError` for errors related to the processing of data/quantities.
+- New module `vibdata` to store vibrational data extracted from input files.
+- New procedure `build_vibdata` to extract vibrational data.
 
 ### Changed